+++ packages/paste-theme/src/generateThemeFromTokens.ts
@@ -1,3 +1,5 @@
+import { sizings as fallbackSizings } from "@twilio-paste/design-tokens/dist/themes/twilio/tokens.es6";
+
 import type {
   BackgroundColorsKeys,
   BorderColorsKeys,
@@ -19,7 +21,7 @@ import type {
 } from "./types/GenericThemeShape";
 import { remToPx } from "./utils/remToPx";
 
-interface GenerateThemeFromTokensArgs {
+export interface GenerateThemeFromTokensArgs {
   backgroundColors: Partial<{ [key in BackgroundColorsKeys]: any }>;
   borderColors: Partial<{ [key in BorderColorsKeys]: any }>;
   borderWidths: Partial<{ [key in BorderWidthsKeys]: any }>;
@@ -56,11 +58,13 @@ export const generateThemeFromTokens = ({
   textColors,
   zIndices,
 }: GenerateThemeFromTokensArgs): GenericThemeShape => {
+  // breakpoints need rm not CSS variables so need to use a fallback for the default sizings
+  const sizingsForBreakpoints = sizings.size0.includes("var") ? fallbackSizings : sizings;
   // default breakpoints
   const breakpoints = [
-    remToPx(sizings.size40, "string"),
-    remToPx(sizings.size100, "string"),
-    remToPx(sizings.size120, "string"),
+    remToPx(sizingsForBreakpoints.size40, "string"),
+    remToPx(sizingsForBreakpoints.size100, "string"),
+    remToPx(sizingsForBreakpoints.size120, "string"),
   ];
 
   return {