+++ packages/paste-website/src/components/color-swatch/ColorGradient.tsx
@@ -63,7 +63,7 @@ export const ColorGradient: React.FC<
 > = ({ aliasPrefix, makeTall = "false", index = 0 }) => {
   const [show, setShow] = React.useState(false);
   const { theme } = useDarkModeContext();
-  const aliasValues = getAliasValuesFromPrefix(aliasPrefix, theme);
+  const aliasValues = getAliasValuesFromPrefix(aliasPrefix, theme || "twilio");
   const count = aliasValues.length - 1;
 
   const { ref, inView } = useInView({